refactor: reorganize component imports and move to features directory
