+++ src/app/(dashboard)/layout.tsx
@@ -1,10 +1,10 @@
-import Sidebar from "@/components/organisms/Sidebar";
+import Sidebar from "@/components/features/dashboard/organisms/Sidebar";
 import React from "react";
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
     return (
         <div className="flex h-screen overflow-hidden">
-            <Sidebar />
+            <Sidebar/>
             <main className="flex-1 overflow-y-auto">
                 {children}
             </main>
