+++ src/components/features/dashboard/templates/DashboardContentLayout.tsx
